Move atom.Labeled to api.Labeled.
Issue: #706
